Update update-smlocate
Stop farting around with the two directory approach.  Handle temp files in destDir instead.

Found an interesting issue with Solaris 9 /bin/sh where "!" is neither an external or a shell keyword.  This broke things like "if [ ! -d something ];", so instead we regress to using ":" e.g. "if [ -d something ]; then ;; else ... fi"

I have tested this with heirloom sh and it breaks on trying to figure out how to use su... so more on that later.